pyproject.toml
[project]
name = "ollama-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"

dependencies = ["ollama", "uvicorn", "python-multipart", "fastapi", "pydantic", "python-docx", "pypdf2"]
